pyproject.toml
[project]
name = "todo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.26",
    "google-generativeai>=0.3.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]
